+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - envs: Register `Cookie-*-*` environments
 - examples: Follow a square patterrn using a base-velocity environment
 - examples: Follow joystick inputs mapped to a base-velocity environment
+- model: Add `CollisionGeometry` dataclass
 - model: Add `JointProperties` class for per-joint simulation properties
 - model: Add `left_wheeled` property
 - model: Add `wheel_base` property